its more correct to use decimals for lat/lng

+++ src/MawDbMigrate/Models/Source/Photo.cs
@@ -38,6 +38,6 @@ public class Photo
     public string? XsSqPath { get; set; }
     public int? XsSqSize { get; set; }
 
-    public double? GpsLatitude { get; set; }
-    public double? GpsLongitude { get; set; }
+    public decimal? GpsLatitude { get; set; }
+    public decimal? GpsLongitude { get; set; }
 }
